updated package

+++ src/SFA.DAS.EmployerCommitmentsV2.Web/Models/Apprentice/Edit/BaseEdit.cs
@@ -132,6 +132,6 @@ namespace SFA.DAS.EmployerCommitmentsV2.Web.Models.Apprentice.Edit
         }
         public int? EmploymentEndMonth { get; set; }
         public int? EmploymentEndYear { get; set; }
-        public decimal? EmploymentPrice { get; set; }
+        public int? EmploymentPrice { get; set; }
     }
 }
